% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gp_impute_core.R
\name{add_MAP_summaries}
\alias{add_MAP_summaries}
\title{Internal \code{ibiVizEdit} utility for appending summary data from an imputation model's posterior distributions}
\usage{
add_MAP_summaries(model_outputs = NULL, pars = c("HR", "R"))
}
\arguments{
\item{model_outputs}{output object generated from the imputation run}

\item{pars}{the parameters from the model for which to generate maximum a posteriori estimates}
}
\value{
appends maximum a posteriori estiamtes from the imputation model to the \code{model_outputs}
}
\description{
Internal \code{ibiVizEdit} utility for appending summary data from an imputation model's posterior distributions
}
