% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_utilities.R
\name{add_ppg_waveform}
\alias{add_ppg_waveform}
\title{Internal \code{ibiVizEdit} utility that adds PPG to plot}
\usage{
add_ppg_waveform(
  base_plot = NULL,
  ppg_data = NULL,
  show_ppg = FALSE,
  time_col = "Time",
  ppg_col = "PPG"
)
}
\description{
Internal \code{ibiVizEdit} utility that adds PPG to plot
}
