% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graphing.R
\name{add_task_v_lines}
\alias{add_task_v_lines}
\title{Internal utility for modifying gui plot}
\usage{
add_task_v_lines(
  base_plot = NULL,
  timing_data = NULL,
  time_col = "Time",
  task_col = "Task",
  label_color = IBI_POINT_COLORS["original"]
)
}
\arguments{
\item{base_plot}{is a \code{ggplot2} object, typically generated by \code{generate_base_gui_plot}}

\item{timing_data}{is a \code{data.frame} that contains time stamps for the start and stop of different
tasks/conditions}

\item{time_col}{is of type \code{character} and is the column name in the \code{timing_data} that contains start/stop
time stamps}

\item{task_col}{is of type \code{character} and is the column name in the \code{timing_data} that contains
task/condition names}

\item{label_color}{hexidecimal code for label used in conjunction with the vertical lines. Defaults to the standard
"original" color}
}
\description{
\code{add_task_v_lines} adds vertical lines that delineate the timing boundaries of tasks/conditions defined in the
timing file
}
