% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{addLogo}
\alias{addLogo}
\title{UI utility for \code{ibiVizEdit} that renders the Dead Reckoning wide Logo}
\usage{
addLogo(logo_filepath = NULL, url = "https://www.deadreckoning.consulting/")
}
\description{
UI utility for \code{ibiVizEdit} that renders the Dead Reckoning wide Logo
}
