% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{addMainFooter}
\alias{addMainFooter}
\title{UI utility for \code{ibiVizEdit} that adds a note about open-source nature of program and link to documentation}
\usage{
addMainFooter(docs_link = NULL, repo_link = NULL, wiki_link = NULL)
}
\arguments{
\item{docs_link}{a hyperlink to additional {ibiVizEdit} documentation}

\item{repo_link}{a hyperlink to the application repository on GH}

\item{wiki_link}{a hyperlink to the Wiki for additional documentation}
}
\description{
UI utility for \code{ibiVizEdit} that adds a note about open-source nature of program and link to documentation
}
