% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{addPpgFooter}
\alias{addPpgFooter}
\title{Utility for \code{ibiVizEdit} that generates footer note at the bottom of the PPG tab}
\usage{
addPpgFooter(heading = "PPG Tab Overview")
}
\description{
Utility for \code{ibiVizEdit} that generates footer note at the bottom of the PPG tab
}
