% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{addProcessingFooter}
\alias{addProcessingFooter}
\title{UI utility for \code{ibiVizEdit} that renders a footer for the processing panel}
\usage{
addProcessingFooter(heading = "Processing Tab Overview")
}
\arguments{
\item{heading}{heading info presented to the user}
}
\description{
UI utility for \code{ibiVizEdit} that renders a footer for the processing panel
}
