% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{AVERAGE_RESPIRATION_BY_AGE}
\alias{AVERAGE_RESPIRATION_BY_AGE}
\title{Prior boundaries based on average respiration rates for each population}
\format{
An object of class \code{list} of length 6.
}
\usage{
AVERAGE_RESPIRATION_BY_AGE
}
\description{
This list is used to define respiration rates and will be incorporated into imputation models to improve accuracy 
and to ensure that a sufficient number of respiration cycles are selected. May be subject to manual override in the 
future but not in the experimental version. Will also be used when HRV values are calculated as part of the final 
set of outputs.
}
\keyword{datasets}
