% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{basic_ppg}
\alias{basic_ppg}
\title{Server-side utility for \code{ibiVizEdit} that dynamically updates pre-processing PPG plot}
\usage{
basic_ppg(ppg_data = NULL, brush_in = NULL)
}
\arguments{
\item{ppg_data}{PPG data series}

\item{brush_in}{the brush used to define the time window for inspection}
}
\description{
Server-side utility for \code{ibiVizEdit} that dynamically updates pre-processing PPG plot
}
