% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{BUTTON_COLORS}
\alias{BUTTON_COLORS}
\title{Vector that specifies three basic color templates based on conditions}
\format{
An object of class \code{character} of length 3.
}
\usage{
BUTTON_COLORS
}
\description{
Vector that specifies three basic color templates based on conditions
}
\keyword{datasets}
