% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{BUTTON_STATUS}
\alias{BUTTON_STATUS}
\title{Initialization of BUTTON_STATUS reactiveValues used throughout the application}
\format{
An object of class \code{reactivevalues} of length 3.
}
\usage{
BUTTON_STATUS
}
\description{
Initialization of BUTTON_STATUS reactiveValues used throughout the application
}
\keyword{datasets}
