% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{click_point_selection}
\alias{click_point_selection}
\title{Server side function that "collects" points for editing when in click ibi selection mode}
\usage{
click_point_selection(
  input,
  click_id,
  dbl_click_id,
  valid_status = "click",
  status_var = reactive({     TEMP_GRAPHICS_SETTINGS[["select_mode"]] })
)
}
\arguments{
\item{input}{{shiny} internal}

\item{click_id}{the click id used to perform point selection in the "main" editing plot on a given panel}

\item{dbl_click_id}{the double click id used to perform point selection in the "main" editing plot on a given panel. 
This action is used to reset or "de-select" any points that were previously highlighted by the user.}

\item{valid_status}{defaults to "click" - the other option is "drag" in terms of point selection}

\item{status_var}{the reactiveValues that "track" whether the select_mode status is "drag" or "click"}
}
\description{
Server side function that "collects" points for editing when in click ibi selection mode
}
