% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{click_ppg_editing}
\alias{click_ppg_editing}
\title{Server side function that enables manual addition and removal of IBI points using the PPG waveform}
\usage{
click_ppg_editing(
  input,
  click_id,
  dbl_click_id,
  valid_status = "edit",
  status_var = reactive({     TEMP_GRAPHICS_SETTINGS[["ppg_mode"]] })
)
}
\arguments{
\item{input}{{shiny} internal}

\item{click_id}{the click id used to perform point selection in the "main" editing plot on a given panel}

\item{dbl_click_id}{the double click id used to perform point selection in the "main" editing plot on a given panel. 
This action is used to reset or "de-select" any points that were previously highlighted by the user.}

\item{valid_status}{defaults to "click" - the other option is "drag" in terms of point selection}

\item{status_var}{the reactiveValues that "track" whether the select_mode status is "drag" or "click"}
}
\description{
Note that this function is mainly intended for use after imputing a section of PPG data. It can however be used
in conjunction with the raw PPG signal. Results may not differ from those produced by the initial peak detection
algorithm because the function leverages the same basic peak detection machinery \(see \code{find_peaks}\), albeit
on a set of points closer to the peak manually identified by the user.
}
