% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{combine_button_action}
\alias{combine_button_action}
\title{Server side function to facilitate combining multiple points}
\usage{
combine_button_action(ibi_data, selected_points = NULL, status = NULL)
}
\arguments{
\item{ibi_data}{IBI data stored in a list of reactiveValues and edited during the user's {ibiVizEdit} session}

\item{selected_points}{the points defined by using the "drag" or "click" selection method}

\item{status}{the status of the combine button - whether it can be used or not}
}
\description{
Server side function to facilitate combining multiple points
}
