% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{create_and_return_gp_output_subdir}
\alias{create_and_return_gp_output_subdir}
\title{Internal \code{ibiVizEdit} utility for creating an subdirectory for GP imputation model outputs}
\usage{
create_and_return_gp_output_subdir(out_dir, gp_driver, sys_time = Sys.time())
}
\arguments{
\item{out_dir}{the output directory set by {create_and_return_output_dir}}

\item{gp_driver}{a \code{list} containing the imputation data and settings}

\item{sys_time}{the timestamp from the system used to differentiate between outputs}
}
\value{
creates and outputs the sub directory for a given Gaussian process imputation run
}
\description{
Internal \code{ibiVizEdit} utility for creating an subdirectory for GP imputation model outputs
}
