% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{create_and_return_output_dir}
\alias{create_and_return_output_dir}
\title{Internal \code{ibiVizEdit} utility for creating output directory inside working directory}
\usage{
create_and_return_output_dir(wd = NULL, case_id = NULL, optional_id = NULL)
}
\description{
Internal \code{ibiVizEdit} utility for creating output directory inside working directory
}
