% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{create_and_return_screenshot_dir}
\alias{create_and_return_screenshot_dir}
\title{Internal \code{ibiVizEdit} utility for creating screenshot directory inside of the output directory}
\usage{
create_and_return_screenshot_dir(out_dir = NULL)
}
\description{
Internal \code{ibiVizEdit} utility for creating screenshot directory inside of the output directory
}
