% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_and_process.R
\name{create_gui_timing_table}
\alias{create_gui_timing_table}
\title{Utility function for transforming timing file inputs into stacked display-friendly table}
\usage{
create_gui_timing_table(df = NULL)
}
\description{
There needs to be a lot of data formatting checks/errors raised here - lots of unit testing as a result too
}
\details{
One check is to make sure that the columns appear arranged in order (may just want a separate format utility to run)
}
