% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{CURRENT_NAME_VERSION}
\alias{CURRENT_NAME_VERSION}
\title{Configuration variable - solely for inclusion in the main header}
\format{
An object of class \code{character} of length 1.
}
\usage{
CURRENT_NAME_VERSION
}
\description{
Configuration variable - solely for inclusion in the main header
}
\keyword{datasets}
