% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{dataEntryActionButtons}
\alias{dataEntryActionButtons}
\title{UI utility for \code{ibiVizEdit} that generates data entry actionButtons}
\usage{
dataEntryActionButtons()
}
\description{
UI utility for \code{ibiVizEdit} that generates data entry actionButtons
}
