% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{divide_button_action}
\alias{divide_button_action}
\title{Server side function to facilitate division of a single point}
\usage{
divide_button_action(
  ibi_data,
  denom = NULL,
  selected_points = NULL,
  status = NULL
)
}
\arguments{
\item{ibi_data}{IBI data stored in a list of reactiveValues and edited during the user's {ibiVizEdit} session}

\item{denom}{the denominator specified by the user. Determines the number of points to divide the selected IBI value
evenly into.}

\item{selected_points}{the points defined by using the "drag" or "click" selection method}

\item{status}{the status of the divide button - whether it can be used or not}
}
\description{
Takes a single point and divides it into n points as determined by the user-specified denominator - defaults to 2 in
the UI
}
