% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{DOCS_LINK}
\alias{DOCS_LINK}
\title{Link to the docs - going to need a significant overhaul}
\format{
An object of class \code{character} of length 1.
}
\usage{
DOCS_LINK
}
\description{
Link to the docs - going to need a significant overhaul
}
\keyword{datasets}
