% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_and_process.R
\name{downsample_ppg_data}
\alias{downsample_ppg_data}
\title{Internal utility that creates a down-sampled data.frame of the ppg signal}
\usage{
downsample_ppg_data(
  ppg_data,
  sampling_rate,
  downsampled_rate = 100,
  ppg_col = "PPG",
  time_col = "Time"
)
}
\description{
Internal utility that creates a down-sampled data.frame of the ppg signal
}
