% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{drag_point_collection}
\alias{drag_point_collection}
\title{Server side function that "collects" points for editing when in click and drag selection mode}
\usage{
drag_point_collection(
  input,
  brush_id,
  valid_status = "drag",
  status_var = reactive({     TEMP_GRAPHICS_SETTINGS[["select_mode"]] })
)
}
\description{
Server side function that "collects" points for editing when in click and drag selection mode
}
