% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{DYNAMIC_DATA}
\alias{DYNAMIC_DATA}
\title{Initialization of DYNAMIC_DATA reactiveValues that can change and be modified based on user actions}
\format{
An object of class \code{reactivevalues} of length 3.
}
\usage{
DYNAMIC_DATA
}
\description{
Initialization of DYNAMIC_DATA reactiveValues that can change and be modified based on user actions
}
\keyword{datasets}
