% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dynamicCheckBoxInput.R
\name{dynamicCheckBoxInputMod}
\alias{dynamicCheckBoxInputMod}
\title{Server Shiny Module for \code{ibiVizEdit} that serves as a basic wrapper and enables dynamic updating of checkbox UI}
\usage{
dynamicCheckBoxInputMod(
  input,
  output,
  session,
  label = NULL,
  choices = NULL,
  selected = NULL,
  paste_char = NULL
)
}
\description{
Server Shiny Module for \code{ibiVizEdit} that serves as a basic wrapper and enables dynamic updating of checkbox UI
}
