% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dynamicClrButton.R
\name{dynamicClrButtonMod}
\alias{dynamicClrButtonMod}
\title{Server module for \code{ibiVizEdit} that dynamically switches actionButton UIs based on color}
\usage{
dynamicClrButtonMod(
  input,
  output,
  session,
  status_name = NULL,
  label = NULL,
  updated_label = NULL,
  default_display_name = NULL,
  button_name = "click_in",
  active_color = BUTTON_COLORS["standard"],
  inactive_color = BUTTON_COLORS["inactive"],
  updated_color = BUTTON_COLORS["warning"]
)
}
\arguments{
\item{input, output, session}{internal {shiny} parameters}

\item{status_name}{a value passed via internal processes of {ibiVizEdit} determining whether a given button is 
available for use.}

\item{label}{the initial label presented on the button rendered in the ui}

\item{updated_label}{the updated if there is a change in label that corresponds with a change in status}

\item{default_display_name}{default name displayed in the {ibiVizEdit} ui}

\item{button_name}{default set to "click_in" used to define the connection between the server and ui modules}

\item{active_color}{the active color - defaults to settings defined in {ibiVizEdit} internals}

\item{inactive_color}{the inactive color - defaults to settings defined in {ibiVizEdit} internals}

\item{updated_color}{the updated color (when active) - defaults to settings defined in {ibiVizEdit} internals}
}
\description{
Server module for \code{ibiVizEdit} that dynamically switches actionButton UIs based on color
}
