% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{dynamicClrButtonModUI}
\alias{dynamicClrButtonModUI}
\title{UI utility for \code{ibiVizEdit} that serves as a basic wrapper and enables single actionButton generation.}
\usage{
dynamicClrButtonModUI(id = NULL, inline = FALSE)
}
\description{
UI utility for \code{ibiVizEdit} that serves as a basic wrapper and enables single actionButton generation.
}
