% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{dynamicNumInputMod}
\alias{dynamicNumInputMod}
\title{Sever-side utility for \code{ibiVizEdit} that dynamically updates data entry options on UI, starts with defaults}
\usage{
dynamicNumInputMod(input, output, session, label = NULL, value = NULL)
}
\description{
Sever-side utility for \code{ibiVizEdit} that dynamically updates data entry options on UI, starts with defaults
}
