% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{dynamicSelectInputMod}
\alias{dynamicSelectInputMod}
\title{Sever-side utility for \code{ibiVizEdit} that dynamically produces a selectInput Field}
\usage{
dynamicSelectInputMod(
  input,
  output,
  session,
  label = NULL,
  choices = NULL,
  choice_index = NULL
)
}
\description{
Sever-side utility for \code{ibiVizEdit} that dynamically produces a selectInput Field
}
