% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{dynamicTextInputMod}
\alias{dynamicTextInputMod}
\title{Server-side utiltiy for \code{ibiVizEdit} that dynamically updates data entry options on UI, starts with defaults}
\usage{
dynamicTextInputMod(input, output, session, label = NULL, value = NULL)
}
\description{
Server-side utiltiy for \code{ibiVizEdit} that dynamically updates data entry options on UI, starts with defaults
}
