% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{EPOCH_CHOICES}
\alias{EPOCH_CHOICES}
\title{Settings for the epochs users can request to have the output "chunked" into}
\format{
An object of class \code{numeric} of length 5.
}
\usage{
EPOCH_CHOICES
}
\description{
Settings for the epochs users can request to have the output "chunked" into
}
\keyword{datasets}
