% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{EPOCH_SELECTED}
\alias{EPOCH_SELECTED}
\title{Defaults for epoch lengths in output files.}
\format{
An object of class \code{numeric} of length 5.
}
\usage{
EPOCH_SELECTED
}
\description{
Defaults for epoch lengths in output files.
}
\keyword{datasets}
