% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{estimate_average_HR}
\alias{estimate_average_HR}
\title{Internal utility for \code{ibiVizEdit} that generates a global estimated mean HR}
\usage{
estimate_average_HR(ibi_data = NULL, ibi_col = "IBI", trim = 3)
}
\arguments{
\item{ibi_data}{\code{data.frame} that contains the IBI time series and corresponding time values}

\item{ibi_col}{name of the column in \code{ibi_data} that contains the IBI values}

\item{trim}{number of seconds to trim from the beginning and end of the IBI series to remove when calculating HR}
}
\value{
estimate of HR in beats per minute
}
\description{
Internal utility for \code{ibiVizEdit} that generates a global estimated mean HR
}
