% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{estimate_avg_respiration}
\alias{estimate_avg_respiration}
\title{Internal utility for determing average respiration jointly using PPG and IBI signals}
\usage{
estimate_avg_respiration(
  ibi_data = NULL,
  respiration_cat = NULL,
  respiration_mapping = AVERAGE_RESPIRATION_BY_AGE,
  ibi_col = "IBI",
  time_col = "Time"
)
}
\arguments{
\item{ibi_data}{\code{data.frame} that contains the IBI time series and corresponding time values}

\item{respiration_cat}{PLACEHOLDER}

\item{respiration_mapping}{PLACEHOLDER}

\item{ibi_col}{name of the column in \code{ibi_data} that contains the IBI values}

\item{time_col}{name of the column in \code{ibi_data} that contains information about the relative}
}
\value{
repiration statistics for the IBI series
}
\description{
Internal utility for determing average respiration jointly using PPG and IBI signals
}
