% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{estimate_max_density}
\alias{estimate_max_density}
\title{Internal utility for estimating mode of posterior distribution from imputation models}
\usage{
estimate_max_density(x)
}
\arguments{
\item{x}{a \code{numeric} series, vector, or matrix of values}
}
\value{
maximum a posteriori value of the input object
}
\description{
\code{estimate_mode} takes the posterior samples from a \code{stanfit} object and returns the mode of the posterior
distribution
}
