% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{extract_ibi_editing_info}
\alias{extract_ibi_editing_info}
\title{Server side function to extract necessary information from the IBI time series}
\usage{
extract_ibi_editing_info(ibi_data, selected_points = NULL)
}
\arguments{
\item{ibi_data}{IBI data stored in a list of reactiveValues and edited during the user's {ibiVizEdit} session}

\item{selected_points}{the points defined by using the "drag" or "click" selection method}
}
\description{
Takes in the ibi time series and extracts information used in the average, combine, and divide computations
}
