% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gp_impute_core.R
\name{extract_valid_local_HP_stats}
\alias{extract_valid_local_HP_stats}
\title{Internal utility for extracting local HP mean and sd based on user-identified valid data}
\usage{
extract_valid_local_HP_stats(
  ibi_data = NULL,
  time_min = NULL,
  time_max = NULL,
  selected_points = NULL,
  input_windows = NULL,
  ibi_col = "IBI",
  time_col = "Time"
)
}
\arguments{
\item{ibi_data}{a \code{data.frame} containing the IBI series and corresponding time}

\item{time_min}{the lower boundary of the imputation window - used to identify the range from which to calculate the 
"local" average heart period}

\item{time_max}{the upper boundary of the imputation window - used to identify the range from which to calculate the
"local" average heart period}

\item{selected_points}{returned from the \code{shiny} brush used to identify the PPG imputation window}

\item{input_windows}{the windows defined by the user and default settings that will be used by the imputation model 
to calculate the replacement PPG data for the targeted selection.}

\item{ibi_col}{the column name in \code{ibi_data} that contains the IBI time series}

\item{time_col}{the column name in the \code{ibi_data} that contains the timing information for the IBI series}
}
\value{
a \code{list} of "local" heart period stats including the mean and standard deviation
}
\description{
Internal utility for extracting local HP mean and sd based on user-identified valid data
}
