% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{FILE_SETTINGS}
\alias{FILE_SETTINGS}
\title{Initialization of FILE_SETTINGS reactiveValues used throughout the application}
\format{
An object of class \code{reactivevalues} of length 3.
}
\usage{
FILE_SETTINGS
}
\description{
Initialization of FILE_SETTINGS reactiveValues used throughout the application
}
\keyword{datasets}
