% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{fileButtons}
\alias{fileButtons}
\title{UI utility for \code{ibiVizEdit} that generates a series of file and director input objects}
\usage{
fileButtons(
  button_color = BUTTON_COLORS["standard"],
  heading = "Select File(s) & Working Directory"
)
}
\description{
UI utility for \code{ibiVizEdit} that generates a series of file and director input objects
}
