% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_input_and_process.R
\name{filter_ppg}
\alias{filter_ppg}
\title{Internal utility that attempts to maximize raw signal properties to generate more reliable peak locations}
\usage{
filter_ppg(ppg_data, sampling_rate, ppg_col = "PPG", time_col = "Time")
}
\description{
Internal utility that attempts to maximize raw signal properties to generate more reliable peak locations
}
