% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_find_ibis.R
\name{find_ibis}
\alias{find_ibis}
\title{Integrated function for generating interbeat interval time-series.}
\usage{
find_ibis(ppg_signal, sampling_rate, min_time, time_adjust = 3, peak_iter)
}
\arguments{
\item{ppg_signal}{a vector of signal values collected from a photopleythysmogram sensor. If not a vector of values 
must be an object that can be safely cast to a single vector of temporally sequenced numeric values.}

\item{sampling_rate}{the desired sampling rate, in Hz, for the data used in the peak detection algorithm.}

\item{min_time}{the minimum time value of the original time series in seconds}

\item{time_adjust}{the time value adjustment applied to the ibi_series. Within the \code{ibiVizEdit} app the value
defaults to 3, which, in effect removes the first and last three seconds of the returned ibi values as a crude 
system for dealing with beginning and end of file issues associated with peak detection.}

\item{peak_iter}{a value that specificies the number of iterations for the algorithm to use when attempting to 
identify the location of the heartbeat "peaks"}
}
\value{
Returns a list containing a \code{dataframe} of interbeat intervals and their timing and a \code{dataframe}
of diagnostic information about the peak detection algorithm optimization parameters and summary statistics.
}
\description{
\code{find_ibis} identifies the timing of heart beats in photoplethysmogram data. Input is a time-series of cardiac
data generated via recordings from photoplethysmogram sensors. The function does three main things: optimizes
bandwidth used in the peak detection algorithm, uses the peaks as proxies for the timing of heart beats, and returns
a time-series of interbeat intervals with basic diagnostic statistics.
}
