% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_find_ibis.R
\name{find_peaks}
\alias{find_peaks}
\title{General peak detection algorithm.}
\usage{
find_peaks(x, bw)
}
\arguments{
\item{x}{a vector of equal interval values measuring a periodic or quasi-periodic process. If not a vector of values,
must be an object that can be safely cast to a single vector of temporally sequenced numeric values.}

\item{bw}{the bandwidth used to define the moving window within which a "peak" is identified.}
}
\value{
Returns the sequentially ordered index values at which each "peak" is detected.
}
\description{
\code{find_peaks} identifies the relative location of maximum values in periodic and quasi-peridodic univariate
time-series. This original function was taken from https://github.com/stas-g/findPeaks/master/find_peaks.R.
}
