% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graphing.R
\name{generate_base_gui_plot}
\alias{generate_base_gui_plot}
\title{Internal utility for gui plot generation}
\usage{
generate_base_gui_plot(
  ibi_data = NULL,
  color_map = NULL,
  ibi_col = "IBI",
  time_col = "Time",
  pnt_type = "pnt_type"
)
}
\arguments{
\item{ibi_data}{is a \code{data.frame} that contains the IBI series being activel edited by the user during the
session}

\item{color_map}{is defined by internal {ibibVizEdit} settings that map point colors to each category}

\item{ibi_col}{is of type \code{character} and is the column name in the \code{ibi_data} that contains the IBI series}

\item{time_col}{is of type \code{character} and is the column name in the \code{ppg_data} that contains the time
variable}

\item{pnt_type}{is the default column name for the IBI data being edited that represents the status of the IBI value
in terms of whether it was originally detected by the peak detection algorithm and if not what type of edit 
generated the final value present in the data set.}
}
\description{
\code{generate_base_gui_plot} defines the base plot that forms the basis for interactively editing IBIs
}
