% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gp_impute_core.R
\name{generate_gp_ppg_predictions}
\alias{generate_gp_ppg_predictions}
\title{Internal \code{ibiVizEdit} utility for selecting maximum a posteriori estimates of corrupted PPG data}
\usage{
generate_gp_ppg_predictions(model_outputs = NULL, gp_driver = NULL)
}
\arguments{
\item{model_outputs}{output object generated from the imputation run}

\item{gp_driver}{a \code{list} object used to define the imputation problem and includes hyperparameters}
}
\value{
output \code{data.frame} that upsamples based on the Gaussian process imputation model
}
\description{
Internal \code{ibiVizEdit} utility for selecting maximum a posteriori estimates of corrupted PPG data
}
