% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{generate_heads_up_info}
\alias{generate_heads_up_info}
\title{Server-side utility for \code{ibiVizEdit} that updates summary stats displayed in text window}
\usage{
generate_heads_up_info(input, hover_id = NULL, ibi_data = NULL)
}
\arguments{
\item{input}{{shiny} internal}

\item{hover_id}{internal name for the hover input object}

\item{ibi_data}{IBI data stored in a list of reactiveValues and edited during the user's {ibiVizEdit} session}
}
\description{
Server-side utility for \code{ibiVizEdit} that updates summary stats displayed in text window
}
