% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_and_process.R
\name{generate_ibis_w_timing}
\alias{generate_ibis_w_timing}
\title{Integrated function for generating IBIs matched with a timing variable}
\usage{
generate_ibis_w_timing(clean_ppg, ppg_col = "PPG", time_col = "Time")
}
\arguments{
\item{clean_ppg}{is the cleaned and processed photoplethysmogram data (either a \code{data.frame} or a
\code{matrix}).}

\item{ppg_col}{is the name of the column in the \code{clean_ppg} object that contains measuresments of the pulse
waveform.}

\item{time_col}{is the name of the column in the \code{clean_ppg} object that measures the passage of time associated
with the pulse waveform.}
}
\value{
Returns a matrix of interbeat intervals and their relative timing.
}
\description{
\code{generate_ibis_w_timing} integrates other \code{ibi_VizEdit} functions \code{iter_IBI} and \code{sum_rev} to
produce a matrix of values. One column stores the returned interbeat intervals. The other column represents a timing
variable. Timing and IBIs are estimated to the nearest millisecond.
}
