% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gp_impute_core.R
\name{generate_model_ppg_inputs}
\alias{generate_model_ppg_inputs}
\title{Internal utility for generating imputation input data}
\usage{
generate_model_ppg_inputs(
  time_min = NULL,
  time_max = NULL,
  ppg_data = NULL,
  total_time = NULL,
  ds = NULL,
  input_windows = NULL,
  ppg_col = "PPG",
  time_col = "Time"
)
}
\arguments{
\item{time_min}{is a \code{numeric} 1D vector that contains the minimum time value of the user-defined imputation
window.}

\item{time_max}{is a \code{numeric} 1D vector that contains the maximum time value of the user-definedimputation
window.}

\item{ppg_data}{is a \code{data.frame} that contains the processed PPG signal and a time variable.}

\item{ds}{is a \code{integer} that represents the final "downsampled" rate of the PPG data, in Hz}

\item{ppg_col}{is of type \code{character} and is the column name in the \code{ppg_data} that contains the PPG signal}

\item{time_col}{is of type \code{character} and is the column name in the \code{ppg_data} that contains the time
variable}

\item{expansion_factor}{is a \code{integer} that represents the number of average respiration cycles the program will
use when exapanding to select input data for the imputation model.}

\item{respiration_cycle_time}{is a \code{numeric} value that indicates the average number of seconds elapsed during
each respiration cycle. The value is derived from \code{estimate_avg_respiration()}}
}
\description{
\code{generate_model_ppg_inputs} is used to generate a \code{data.frame} that contains an appropriately down-sampled
set of "Time" and "PPG" inputs for the imputation model. These inputs represent PPG data that "surrounds" the window
selected for imputation. The size of the window is dynamically determined for each participant based on average
respiration rate (derived from \code{estimate_avg_respiration()}).
}
