% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{generate_path_messages}
\alias{generate_path_messages}
\title{Server-side utility for \code{ibiVizEdit} that outputs relevant file and directory information when submitted}
\usage{
generate_path_messages(default_text = NULL, msg_part1 = NULL, obj_name = NULL)
}
\description{
Server-side utility for \code{ibiVizEdit} that outputs relevant file and directory information when submitted
}
