% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_utilities.R
\name{generate_ppg_data_check_plot}
\alias{generate_ppg_data_check_plot}
\title{Internal utility for plot generation}
\usage{
generate_ppg_data_check_plot(
  ppg_data = NULL,
  ppg_col = "PPG",
  time_col = "Time"
)
}
\arguments{
\item{ppg_data}{is a \code{data.frame} that contains the processed PPG signal and a time variable.}

\item{ppg_col}{is of type \code{character} and is the column name in the \code{ppg_data} that contains the PPG signal}

\item{time_col}{is of type \code{character} and is the column name in the \code{ppg_data} that contains the time
variable}
}
\description{
\code{generate_ppg_data_check_plot} defines the properties of the plot called on the UI Processing Panel to check
that the PPG data loading process was successful. The utility and resulting plot allow a user to visually inspect
the imported signal based on settings defined on the UI Data Entry panel
}
