% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{get_user_folder}
\alias{get_user_folder}
\title{Finding user home directory.}
\usage{
get_user_folder(input)
}
\description{
\code{get_user_folder} is designed to identify the home directory of the user who initiated the R session.
}
