% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{get_working_directory}
\alias{get_working_directory}
\title{Sever-side utility for \code{ibiVizEdit} that observes and updates directory information}
\usage{
get_working_directory(input, input_name = NULL)
}
\description{
Sever-side utility for \code{ibiVizEdit} that observes and updates directory information
}
