% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gp_impute_Bayes.R
\name{gp_impute_driver}
\alias{gp_impute_driver}
\title{Internal utility that aggregates data and settings needed to trigger an imputation model run}
\usage{
gp_impute_driver(
  iter = NULL,
  warmup = NULL,
  adapt_delta = NULL,
  time_min = NULL,
  time_max = NULL,
  ppg_data = NULL,
  ibi_data = NULL,
  respiration_cat = NULL,
  ds = NULL,
  selected_ibis = NULL,
  ppg_col = "PPG",
  ibi_col = "IBI",
  time_col = "Time",
  expansion_factor = 3
)
}
\arguments{
\item{iter}{number of iterations per chain}

\item{warmup}{number of warmup iterations}

\item{adapt_delta}{parameter passed to rstan call}

\item{time_min}{lower boundary of the time window for imputation}

\item{time_max}{upper boundary of the time window for imputation}

\item{ppg_data}{the processed PPG data.frame be used in the analysis}

\item{ibi_data}{the IBI data used to define priors for the imputation model}

\item{respiration_cat}{the population age group used to define the respiration range}

\item{ds}{the downsampled sampling rate for the PPG data}

\item{selected_ibis}{IBI values that will be used in setting the priors (defined by a combination of the expansion
factor and the individual respiration rate.}

\item{ppg_col}{column name in \code{ppg_data} that contains the PPG signal data. Default is "PPG".}

\item{ibi_col}{column name in \code{ibi_data} that contains the IBI series}

\item{time_col}{defaults to "Time" and represents the synchronized time values for the \code{ppg_data} and 
\code{ibi_data}}

\item{expansion_factor}{is the number of respiration cycles to use in setting the data used to define the paramters
in the imputation model}
}
\value{
applies a series of helper function and returns a \code{list} of values that are used to define the Gaussian
process imputation model.
}
\description{
This could use a considerable refactor - need to back through and break up a lot of these complicated functions
}
