% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graphing.R
\name{highlight_ibis}
\alias{highlight_ibis}
\title{Internal \code{ibiVizEdit} utility that enables dynamic color selection based on selected points}
\usage{
highlight_ibis(
  base_plot = NULL,
  selected_points = NULL,
  time_col = "Time",
  ibi_col = "IBI"
)
}
\description{
Internal \code{ibiVizEdit} utility that enables dynamic color selection based on selected points
}
