% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graphing.R
\name{hover_point_selection}
\alias{hover_point_selection}
\title{Server side function to acquire hover points}
\usage{
hover_point_selection(input, hover_id)
}
\description{
Server side function to acquire hover points
}
