% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{IBI_POINT_COLORS}
\alias{IBI_POINT_COLORS}
\title{Central location for IBI point colors used throughout the plots based on point type as defined in this named vector}
\format{
An object of class \code{character} of length 5.
}
\usage{
IBI_POINT_COLORS
}
\description{
Central location for IBI point colors used throughout the plots based on point type as defined in this named vector
}
\keyword{datasets}
