% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graphing.R
\name{ibi_value_label}
\alias{ibi_value_label}
\title{Internal utility for generating dynamic text label corresponding to ibi value}
\usage{
ibi_value_label(
  base_plot = NULL,
  hover_point = NULL,
  time_col = "Time",
  ibi_col = "IBI"
)
}
\description{
Internal utility for generating dynamic text label corresponding to ibi value
}
