% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{ibiEditingPlots}
\alias{ibiEditingPlots}
\title{Utility for \code{ibiVizEdit} that generates main editing plots for IBI tab}
\usage{
ibiEditingPlots()
}
\description{
Utility for \code{ibiVizEdit} that generates main editing plots for IBI tab
}
