% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{ibiEditingTools}
\alias{ibiEditingTools}
\title{Utility for \code{ibiVizEdit} that gemerates UI components for ibi editing tab tools}
\usage{
ibiEditingTools()
}
\description{
Utility for \code{ibiVizEdit} that gemerates UI components for ibi editing tab tools
}
