% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{IMPUTE_DATA}
\alias{IMPUTE_DATA}
\title{Imputation data - should be ephemeral though certain properties will be appended to output/logs in the end}
\format{
An object of class \code{reactivevalues} of length 3.
}
\usage{
IMPUTE_DATA
}
\description{
Imputation data - should be ephemeral though certain properties will be appended to output/logs in the end
}
\keyword{datasets}
