% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{ini_global_reactive_values}
\alias{ini_global_reactive_values}
\title{Initialize global reactive values.}
\usage{
ini_global_reactive_values()
}
\arguments{
\item{tot_edits}{is a \code{dataframe} that is used to maintain a running track record unique edits made by a user.}

\item{base_on}{is used to toggle access to the "base" editing functions on and off. Default is off.}

\item{adv_on}{is used to toggle access to the "advanced" editing functions on and off. Default is off.}

\item{ppg_on}{is used to toggle the display of the underlying PPG waveform on and off. Only works when a user is
zoomed in on a section of the file. Default is off.}

\item{select_on}{is used to toggle on access to the select brush on the "base" editing panel. This particular select
brush supports the use of the "combine", "divide", and "average" functions. Default is off.}

\item{add_delete_on}{is used to toggle on access to mouse-click-based "add" and "delete" functionality.}

\item{select_on2}{is used to toggle on access to the select brush on the "advanced" editing panel. This particular
select brush allows the user to specify the section of corrupted data targeted for imputation. Default is off.}

\item{add_delete_on2}{is used to toggle on access to mouse-click-based "add" and "delete" functionality. Default is
off.}

\item{start.time}{is used to store the time and date the user began processing a target file.}

\item{GP_impute_tab}{is used to store ongoing information about Gaussian imputation attempts. The table is included
in the final output.}
}
\description{
\code{ini_global_reactive_values} is an internal utility that initializes a set of global reactive values that affect
the behavior of the GUI, including colors, displays, and which actions are available to a user.
}
