% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_gui_editor.R
\name{launch_ibiVizEdit}
\alias{launch_ibiVizEdit}
\title{Utility function for launching \code{ibiVizEdit} RShiny graphic user interface}
\usage{
launch_ibiVizEdit()
}
\description{
Utility function for launching \code{ibiVizEdit} RShiny graphic user interface
}
