% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{LOAD_BUTTON_MESSAGE}
\alias{LOAD_BUTTON_MESSAGE}
\title{Message to display underneath "Load" button}
\format{
An object of class \code{character} of length 1.
}
\usage{
LOAD_BUTTON_MESSAGE
}
\description{
Message to display underneath "Load" button
}
\keyword{datasets}
