% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{load_files_and_settings}
\alias{load_files_and_settings}
\title{Server-side utility for \code{ibiVizEdit} that loads data using specifications set in Data Entry tab}
\usage{
load_files_and_settings(input)
}
\description{
Server-side utility for \code{ibiVizEdit} that loads data using specifications set in Data Entry tab
}
