% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_and_process.R
\name{load_ppg}
\alias{load_ppg}
\title{Internal utility for loading raw photoplethysmogram (PPG) data}
\usage{
load_ppg(
  file_name = NULL,
  skip_lines = NULL,
  column = NULL,
  sampling_rate = NULL
)
}
\arguments{
\item{file_name}{is of type \code{character} and identifies the file path containing the raw PPG signal.}

\item{skip_lines}{is of type \code{int} and specifies the number of lines in the raw file to skip.}

\item{column}{is of type \code{int} and specifies the column containing in the raw file containing the desired PPG
signal.}

\item{sampling_rate}{is of type \code{int} and specifies the hardware sampling rate for the raw signal in Hz}
}
\value{
returns a data.frame that contains the pre-processed PPG signal and a corresponding set of time codes,
initialized at 0.
}
\description{
\code{load_ppg_data} loads the selected raw PPG data for subsequent processing and editing. Accepts tab-delimited
.txt formats only.
}
