% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_and_process.R
\name{load_timing_data}
\alias{load_timing_data}
\title{Internal utility for loading timing file}
\usage{
load_timing_data(file_name = NULL, case_id = NULL)
}
\arguments{
\item{file_name}{is of type \code{character} and represents the filepath containing timing information for the PPG
file of interest, with the first column being a unique identifier.}

\item{case_id}{is of type \code{character} and represents the unique identifier for the target PPG file. The value
is used to match the correct timing information to the target PPG file.}
}
\value{
returns a vector of time stamps that correspond to the start and stop time codes for tasks/conditions of
interest
}
\description{
\code{load_timing_data} loads the selected timing information for the targeted file. The timing file must be
formatted in such a way that the first column represents the file ID, and each column matches the start and stop of
sequential tasks/conditions of interest. Accepts tab-delimited .txt or .csv formats.
}
