% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{META_DATA}
\alias{META_DATA}
\title{Initialization of META_DATA reactiveValues used throughout the application}
\format{
An object of class \code{reactivevalues} of length 3.
}
\usage{
META_DATA
}
\description{
Initialization of META_DATA reactiveValues used throughout the application
}
\keyword{datasets}
