% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{optionalSettingsEntry}
\alias{optionalSettingsEntry}
\title{UI utility for \code{ibiVizEdit} that generates data entry fields for optional settings}
\usage{
optionalSettingsEntry(heading = "Optional Settings:")
}
\description{
UI utility for \code{ibiVizEdit} that generates data entry fields for optional settings
}
