% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_utils.R
\name{out_text_dir}
\alias{out_text_dir}
\title{Shiny reactive object.}
\usage{
out_text_dir()
}
\description{
\code{out_text_dir} is an internal utility that prints either a warning that a working directory has not yet been
selected or the file path of the working directory selected directly in the UI.
}
