% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_utils.R
\name{out_text_ppg_file}
\alias{out_text_ppg_file}
\title{Shiny reactive object.}
\usage{
out_text_ppg_file()
}
\description{
\code{out_text_ppg_file} is an internal utility that prints either a warning that a raw PPG file has not yet been
selected or the file path of the PPG file selected directly in the UI.
}
