% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_utils.R
\name{out_text_time_file}
\alias{out_text_time_file}
\title{Shiny reactive object.}
\usage{
out_text_time_file()
}
\description{
\code{out_text_time_file} is an internal utility prints either a warning that a raw timing file has not yet been
selected or the file path of the timing file directly in the UI.
}
