% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graphing.R
\name{ppg_data_check_empty_plot}
\alias{ppg_data_check_empty_plot}
\title{Internal utility for defualt plot message generation}
\usage{
ppg_data_check_empty_plot()
}
\description{
\code{ppg_data_check_empty_plot} creates a plot with a user-facing message stating that no data has been loaded.
}
