% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{ppg_editing_plot}
\alias{ppg_editing_plot}
\title{Server-side utility for \code{ibiVizEdit} that defines main PPG plot for GUI editing}
\usage{
ppg_editing_plot(ibi_data = DYNAMIC_DATA[["edited_ibi"]], brush_in = NULL)
}
\arguments{
\item{ibi_data}{IBI data stored in a list of reactiveValues and edited during the user's {ibiVizEdit} session}

\item{brush_in}{the brush used to define the time window for inspection}
}
\description{
Server-side utility for \code{ibiVizEdit} that defines main PPG plot for GUI editing
}
