% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{ppgDataPropertiesEntry}
\alias{ppgDataPropertiesEntry}
\title{UI utility for \code{ibiVizEdit} that generates data entry fields for data properties; accepts user-defined defaults}
\usage{
ppgDataPropertiesEntry(heading = "Data Properties:")
}
\description{
UI utility for \code{ibiVizEdit} that generates data entry fields for data properties; accepts user-defined defaults
}
