% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{preProcessPlots}
\alias{preProcessPlots}
\title{Utility for \code{ibiVizEdit} that defines plots for pre-processing tab}
\usage{
preProcessPlots(heading = "Visualize Pre-Processed PPG Data:")
}
\arguments{
\item{heading}{heading info presented to the user}
}
\description{
Utility for \code{ibiVizEdit} that defines plots for pre-processing tab
}
