% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{preProcessTables}
\alias{preProcessTables}
\title{Utility for \code{ibiVizEdit} that defines tables for pre-processing tab}
\usage{
preProcessTables(heading = "Task Timing and Peak Detection Outputs:")
}
\description{
Utility for \code{ibiVizEdit} that defines tables for pre-processing tab
}
