% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{PROCESSING_DEFAULTS}
\alias{PROCESSING_DEFAULTS}
\title{Presets for the inputs}
\format{
An object of class \code{list} of length 9.
}
\usage{
PROCESSING_DEFAULTS
}
\description{
Presets for the inputs
}
\keyword{datasets}
