% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{processing_settings_observer}
\alias{processing_settings_observer}
\title{Server-side utility for \code{ibiVizEdit} that monitors data entry values and updates them accordingly}
\usage{
processing_settings_observer(input_name)
}
\description{
Server-side utility for \code{ibiVizEdit} that monitors data entry values and updates them accordingly
}
