% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{raise_not_in_range_integer}
\alias{raise_not_in_range_integer}
\title{Internal \code{ibiVizEdit} utility that generates a warning if an input expected to be an integer is not}
\usage{
raise_not_in_range_integer(
  input_val = NULL,
  input_name = NULL,
  lower_bound = NULL,
  upper_bound = NULL
)
}
\arguments{
\item{input_val}{the value passed by the user to the {ibiVizEdit} gui}

\item{input_name}{the name corresponding to the {shiny} {input} being evaluated}

\item{lower_bound}{the lower bound of the allowable integer range}

\item{upper_bound}{the upper bound of the allowable integer range}
}
\value{
raises a warning if value is not an integer or not in permitted range
}
\description{
Internal \code{ibiVizEdit} utility that generates a warning if an input expected to be an integer is not
}
