% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{range01}
\alias{range01}
\title{Set min/max to 0/1.}
\usage{
range01(x)
}
\description{
\code{range01} is an internal utility that takes a vector of values and returns a new vector with the same
distribution, re-scaled to a range of 0 to 1.
}
