% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gp_impute_Bayes.R
\name{replace_w_imputed}
\alias{replace_w_imputed}
\title{Internal \code{ibiVizEdit} function for replacing corrupted data with imputed data}
\usage{
replace_w_imputed(
  ppg_out = NULL,
  model_outputs = NULL,
  ppg_col = "PPG",
  time_col = "Time"
)
}
\description{
Internal \code{ibiVizEdit} function for replacing corrupted data with imputed data
}
