% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{REPO_LINK}
\alias{REPO_LINK}
\title{Link to the repo}
\format{
An object of class \code{character} of length 1.
}
\usage{
REPO_LINK
}
\description{
Link to the repo
}
\keyword{datasets}
