% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{RESET_ALL_MESSAGE}
\alias{RESET_ALL_MESSAGE}
\title{Message to display underneath "Reset" button}
\format{
An object of class \code{character} of length 1.
}
\usage{
RESET_ALL_MESSAGE
}
\description{
Message to display underneath "Reset" button
}
\keyword{datasets}
