% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{restore_button_action}
\alias{restore_button_action}
\title{Server side utility that takes restores all IBIs within the selected window}
\usage{
restore_button_action(
  input,
  restore_id,
  edited_data,
  original_data,
  brush_id,
  ibi_or_ppg = NULL
)
}
\arguments{
\item{input}{{shiny} internal}

\item{restore_id}{id value for the button used to enable the restore action}

\item{edited_data}{data stored in a list of reactiveValues that contains "in-progress" edits during the user session}

\item{original_data}{data stored in a list of reactiveValues that contains the "original" data as it existed 
following preliminary processing but before manual editing.}

\item{brush_id}{id for the brush that defines the section of data to restore}

\item{ibi_or_ppg}{used to define whether the data being restored is from the IBI or PPG series}
}
\description{
Server side utility that takes restores all IBIs within the selected window
}
