% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gp_impute_Bayes.R
\name{run_bayesian_gp}
\alias{run_bayesian_gp}
\title{Internal \code{ibiVizEdit} utility for running a Bayesian Gaussian process imputation model}
\usage{
run_bayesian_gp(gp_driver, imputation_model)
}
\arguments{
\item{gp_driver}{\code{list} returned from \code{ibiVizEdit::gp_impute_driver} that includes all of the values 
required for the imputation model.}

\item{imputation_model}{a string that defines a \code{Stan} model}
}
\value{
Returns the outputs of a Bayesian imputation model as defined by \code{imputation_model}
}
\description{
Internal \code{ibiVizEdit} utility for running a Bayesian Gaussian process imputation model
}
