% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_impute_utils.R
\name{save_gp_imputation_traceplots}
\alias{save_gp_imputation_traceplots}
\title{Internal \code{ibiVizEdit} utility for generating diagnostic traceplots}
\usage{
save_gp_imputation_traceplots(
  model_outputs = NULL,
  gp_driver = NULL,
  sub_id = NULL,
  secondary_id = NULL,
  study_id = NULL,
  out_dir = NULL
)
}
\description{
Internal \code{ibiVizEdit} utility for generating diagnostic traceplots
}
