% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{SAVE_OUT_BUTTON_MESSAGE}
\alias{SAVE_OUT_BUTTON_MESSAGE}
\title{Message to display underneath "Save and Output" button}
\format{
An object of class \code{character} of length 1.
}
\usage{
SAVE_OUT_BUTTON_MESSAGE
}
\description{
Message to display underneath "Save and Output" button
}
\keyword{datasets}
