% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graphing.R
\name{save_screenshot}
\alias{save_screenshot}
\title{Server side function that takes and saves screenshot in a screenshots folder for easy sharing with colleagues}
\usage{
save_screenshot(input, data, time_brush, button_id, ibi_or_ppg = NULL)
}
\arguments{
\item{input}{{shiny} internal}

\item{data}{data used to create the plot}

\item{time_brush}{brush that defines the time window displayed in the ui for the user to edit}

\item{button_id}{id value for the button that is intended to "trigger" the creation of the plot}

\item{ibi_or_ppg}{PLACEHOLDER}
}
\value{
saves version of the plot currently being displayed. The intent is to make it easier to share out with 
teammates when editing difficult sections of data and to document pre and post processing efforts
}
\description{
Server side function that takes and saves screenshot in a screenshots folder for easy sharing with colleagues
}
