% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{STATIC_DATA}
\alias{STATIC_DATA}
\title{Initialization of STATIC_DATA reactiveValues that, once processed by the application never change}
\format{
An object of class \code{reactivevalues} of length 3.
}
\usage{
STATIC_DATA
}
\description{
Initialization of STATIC_DATA reactiveValues that, once processed by the application never change
}
\keyword{datasets}
