% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{store_raw_data_filepath}
\alias{store_raw_data_filepath}
\title{Server-side utility for \code{ibiVizEdit} that observes and updates raw data paths file paths}
\usage{
store_raw_data_filepath(input, input_name = NULL)
}
\arguments{
\item{input, input_name}{{shiny} and {ibiVizEdit} internals for defining raw data path}
}
\description{
Server-side utility for \code{ibiVizEdit} that observes and updates raw data paths file paths
}
