% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_find_ibis.R
\name{sum_rev}
\alias{sum_rev}
\title{Internal utility for creating IBI time-series.}
\usage{
sum_rev(x)
}
\arguments{
\item{x}{a vector of equal interval values measuring a periodic or quasi-periodic process. If not a vector of values,
must be an object that can be safely cast to a single vector of temporally sequenced numeric values.}
}
\value{
Returns the sequentially ordered index values at which each "peak" is detected.
}
\description{
\code{sum_rev}
}
