% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{SUMMARY_STATS}
\alias{SUMMARY_STATS}
\title{Initialization of SUMMARY_STATS reactiveValues used to store summary values calculated during application usage}
\format{
An object of class \code{reactivevalues} of length 3.
}
\usage{
SUMMARY_STATS
}
\description{
Initialization of SUMMARY_STATS reactiveValues used to store summary values calculated during application usage
}
\keyword{datasets}
