% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{TAG_LINE}
\alias{TAG_LINE}
\title{Configuration setting - Tagline for the application header}
\format{
An object of class \code{character} of length 1.
}
\usage{
TAG_LINE
}
\description{
Configuration setting - Tagline for the application header
}
\keyword{datasets}
