% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{TEMP_GRAPHICS_SETTINGS}
\alias{TEMP_GRAPHICS_SETTINGS}
\title{Initialization of TEMP_GRAPHICS_SETTINGS reactiveValues that specify the default settings for certain graphics. All 
can be modified by user when interacting with the application.}
\format{
An object of class \code{reactivevalues} of length 3.
}
\usage{
TEMP_GRAPHICS_SETTINGS
}
\description{
Initialization of TEMP_GRAPHICS_SETTINGS reactiveValues that specify the default settings for certain graphics. All 
can be modified by user when interacting with the application.
}
\keyword{datasets}
