% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{THUMB_LOGO}
\alias{THUMB_LOGO}
\title{Location of the thumbnail - used mainly as the favicon}
\format{
An object of class \code{character} of length 1.
}
\usage{
THUMB_LOGO
}
\description{
Location of the thumbnail - used mainly as the favicon
}
\keyword{datasets}
