% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_and_process.R
\name{time_center}
\alias{time_center}
\title{Internal utility for ibiVizEdit that centers PPG and IBI timing values a the start of a task file.}
\usage{
time_center(x, time_col = "Time", time_file = NULL)
}
\arguments{
\item{x}{description}

\item{time_col}{descirption}

\item{time_file}{description}
}
\value{
Returns a re-coded version of the inputs with new timing
}
\description{
\code{time_center} DESCRIPTION
}
