% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_find_ibis.R
\name{time_diff}
\alias{time_diff}
\title{Internal utility that creates IBI values from the find_peaks output}
\usage{
time_diff(x)
}
\arguments{
\item{x}{a vector of time values used to generate a diff series. Time values represent the moment at which a beat was
detected}
}
\value{
Returns a series of differenced values representing the time between each successive x value from the input
series
}
\description{
Internal utility that creates IBI values from the find_peaks output
}
