% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_find_ibis.R
\name{time_sum}
\alias{time_sum}
\title{Internal utility that creates IBI values from the find_peaks output}
\usage{
time_sum(x)
}
\description{
Internal utility that creates IBI values from the find_peaks output
}
