% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{track_editing_options}
\alias{track_editing_options}
\title{Server side function that tracks and updates the status of the editing functions}
\usage{
track_editing_options()
}
\description{
Server side function that tracks and updates the status of the editing functions
}
