% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_hotkey.R
\name{track_hotkey_presses}
\alias{track_hotkey_presses}
\title{Internal utility to enable hotkey functionality during an editing session}
\usage{
track_hotkey_presses(key = NULL, key_map = NULL, button_name = NULL)
}
\description{
see additional examples here:
https://stackoverflow.com/questions/32002170/r-shiny-enabling-keyboard-shortcuts
and here:
https://stackoverflow.com/questions/44500617/r-shiny-key-and-actionbutton-binding-to-reactive-values
}
