% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{TRIGGERS}
\alias{TRIGGERS}
\title{Initiatlization of a set of TRIGGERS reactiveValues used in conjunction with the eventTrigger module}
\format{
An object of class \code{reactivevalues} of length 3.
}
\usage{
TRIGGERS
}
\description{
Initiatlization of a set of TRIGGERS reactiveValues used in conjunction with the eventTrigger module
}
\keyword{datasets}
