% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_input_and_process.R
\name{trim_ppg_window}
\alias{trim_ppg_window}
\title{Internal utility that trims down the time range of the PPG signal used to edit the data}
\usage{
trim_ppg_window(ppg_data, timing_data, time_col = "Time")
}
\arguments{
\item{ppg_data}{\code{data.frame} that contains the processed PPG data}

\item{timing_data}{\code{data.frame} user defined input data set containing task timing}

\item{time_col}{column name in \code{ppg_data} that contains the timing information corresponding with the PPG
signal}
}
\value{
a \code{data.frame} that has been restricted to the range of the timing file +/- 3 seconds
}
\description{
Internal utility that trims down the time range of the PPG signal used to edit the data
}
