% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{user_action_log}
\alias{user_action_log}
\title{Internal \code{ibiVizEdit} utility that creates a running log of user actions}
\usage{
user_action_log(button_name)
}
\description{
Internal \code{ibiVizEdit} utility that creates a running log of user actions
}
