% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{WIDE_LOGO}
\alias{WIDE_LOGO}
\title{Location of the main Dead Reckoning logo}
\format{
An object of class \code{character} of length 1.
}
\usage{
WIDE_LOGO
}
\description{
Location of the main Dead Reckoning logo
}
\keyword{datasets}
