% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init_global_vars.R
\docType{data}
\name{WIKI_LINK}
\alias{WIKI_LINK}
\title{Link to the wiki - needs to be considerably added to}
\format{
An object of class \code{character} of length 1.
}
\usage{
WIKI_LINK
}
\description{
Link to the wiki - needs to be considerably added to
}
\keyword{datasets}
