% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_impute_utils.R
\name{add_MAP_summaries}
\alias{add_MAP_summaries}
\title{Internal \code{ibiVizEdit} utility for appending summary data from an imputation model's posterior distributions}
\usage{
add_MAP_summaries(model_outputs = NULL, pars = c("HR", "R"))
}
\description{
Internal \code{ibiVizEdit} utility for appending summary data from an imputation model's posterior distributions
}
