% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{addIbiFooter}
\alias{addIbiFooter}
\title{Utility for \code{ibiVizEdit} that generates footer note at the bottom of the IBI tab}
\usage{
addIbiFooter(heading = "IBI Tab Overview")
}
\description{
Utility for \code{ibiVizEdit} that generates footer note at the bottom of the IBI tab
}
