% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{addLogo}
\alias{addLogo}
\title{UI utility for \code{ibiVizEdit} that renders the Dead Reckoning wide Logo}
\usage{
addLogo(logo_filepath = NULL, url = "https://www.deadreckoning.consulting/")
}
\arguments{
\item{logo_filepath}{relative path for the desired logo}

\item{url}{the url the user should be directed to upon clicking the logo}
}
\description{
UI utility for \code{ibiVizEdit} that renders the Dead Reckoning wide Logo
}
