% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{addPpgFooter}
\alias{addPpgFooter}
\title{Utility for \code{ibiVizEdit} that generates footer note at the bottom of the PPG tab}
\usage{
addPpgFooter(heading = "PPG Tab Overview")
}
\arguments{
\item{heading}{heading info presented to the user}
}
\description{
Utility for \code{ibiVizEdit} that generates footer note at the bottom of the PPG tab
}
