% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{average_button_action}
\alias{average_button_action}
\title{Server side function to facilitate averaging multiple points}
\usage{
average_button_action(ibi_data, selected_points = NULL, status = NULL)
}
\description{
Takes a single point and divides it into n points as determined by the user-specified denominator - defaults to 2 in
the UI
}
