% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{basic_ppg}
\alias{basic_ppg}
\title{Server-side utility for \code{ibiVizEdit} that dynamically updates pre-processing PPG plot}
\usage{
basic_ppg(ppg_data = NULL, brush_in = NULL)
}
\description{
Server-side utility for \code{ibiVizEdit} that dynamically updates pre-processing PPG plot
}
