% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gp_impute_core.R
\name{check_for_large_rhats}
\alias{check_for_large_rhats}
\title{Internal \code{ibiVizEdit} utility for identying large Rhat estimates in GP imputation model outputs}
\usage{
check_for_large_rhats(summary_table = NULL)
}
\description{
Internal \code{ibiVizEdit} utility for identying large Rhat estimates in GP imputation model outputs
}
