% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{click_point_selection}
\alias{click_point_selection}
\title{Server side function that "collects" points for editing when in click ibi selection mode}
\usage{
click_point_selection(
  input,
  click_id,
  dbl_click_id,
  valid_status = "click",
  status_var = reactive({     TEMP_GRAPHICS_SETTINGS[["select_mode"]] })
)
}
\description{
Server side function that "collects" points for editing when in click ibi selection mode
}
