% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{combine_button_action}
\alias{combine_button_action}
\title{Server side function to facilitate combining multiple points}
\usage{
combine_button_action(ibi_data, selected_points = NULL, status = NULL)
}
\description{
Server side function to facilitate combining multiple points
}
