% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{create_and_return_gp_output_subdir}
\alias{create_and_return_gp_output_subdir}
\title{Internal \code{ibiVizEdit} utility for creating an subdirectory for GP imputation model outputs}
\usage{
create_and_return_gp_output_subdir(out_dir, gp_driver, sys_time = Sys.time())
}
\description{
Internal \code{ibiVizEdit} utility for creating an subdirectory for GP imputation model outputs
}
