% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{create_and_return_gp_output_subdir}
\alias{create_and_return_gp_output_subdir}
\title{Internal \code{ibiVizEdit} utility for creating an subdirectory for GP imputation model outputs}
\usage{
create_and_return_gp_output_subdir(out_dir, gp_driver)
}
\description{
Internal \code{ibiVizEdit} utility for creating an subdirectory for GP imputation model outputs
}
