% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{create_and_return_output_dir}
\alias{create_and_return_output_dir}
\title{Internal \code{ibiVizEdit} utility for creating output directory inside working directory}
\usage{
create_and_return_output_dir(wd = NULL, case_id = NULL, optional_id = NULL)
}
\arguments{
\item{wd}{the working directory for the project}

\item{case_id}{id used by {ibiVizEdit} to track outputs}

\item{optional_id}{id value that can be provided by the user to further differentiate files and directories}
}
\value{
the output directory which combines the file settings. Creates the directory if not present.
}
\description{
Internal \code{ibiVizEdit} utility for creating output directory inside working directory
}
