% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{create_and_return_screenshot_dir}
\alias{create_and_return_screenshot_dir}
\title{Internal \code{ibiVizEdit} utility for creating screenshot directory inside of the output directory}
\usage{
create_and_return_screenshot_dir(out_dir = NULL)
}
\arguments{
\item{out_dir}{the output directory set by {create_and_return_output_dir}}
}
\value{
a directory for saving "screenshots" from {ibiVizEdit}
}
\description{
Internal \code{ibiVizEdit} utility for creating screenshot directory inside of the output directory
}
