% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_input_and_process.R
\name{create_gui_timing_table}
\alias{create_gui_timing_table}
\title{Utility function for transforming timing file inputs into stacked display-friendly table}
\usage{
create_gui_timing_table(df = NULL)
}
\arguments{
\item{df}{the original timing data}
}
\value{
\code{data.frame} formatted for a clean display on the processing tab in the ibiVizEdit \code{RShiny} gui
}
\description{
There needs to be a lot of data formatting checks/errors raised here - lots of unit testing as a result too
}
