% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{divide_button_action}
\alias{divide_button_action}
\title{Server side function to facilitate division of a single point}
\usage{
divide_button_action(
  ibi_data,
  denom = NULL,
  selected_points = NULL,
  status = NULL
)
}
\description{
Takes a single point and divides it into n points as determined by the user-specified denominator - defaults to 2 in
the UI
}
