% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{drag_point_collection}
\alias{drag_point_collection}
\title{Server side function that "collects" points for editing when in click and drag selection mode}
\usage{
drag_point_collection(
  input,
  brush_id,
  valid_status = "drag",
  status_var = reactive({     TEMP_GRAPHICS_SETTINGS[["select_mode"]] })
)
}
\arguments{
\item{input}{{shiny} internal}

\item{brush_id}{the brush id used to perform point selection in the "main" editing plot on a given panel}

\item{valid_status}{defaults to "drag" - the other option is "click" in terms of point selection}

\item{status_var}{the reactiveValues that "track" whether the select_mode status is "drag" or "click"}
}
\description{
Server side function that "collects" points for editing when in click and drag selection mode
}
