% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{dynamicCheckBoxInputMod}
\alias{dynamicCheckBoxInputMod}
\title{Sever-side utility for \code{ibiVizEdit} that dynamically produces an awesomeCheckboxGroup with pre-selections}
\usage{
dynamicCheckBoxInputMod(
  input,
  output,
  session,
  label = NULL,
  choices = NULL,
  selected = NULL,
  paste_char = NULL
)
}
\description{
Sever-side utility for \code{ibiVizEdit} that dynamically produces an awesomeCheckboxGroup with pre-selections
}
