% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dynamicCheckBoxInput.R
\name{dynamicCheckBoxInputMod}
\alias{dynamicCheckBoxInputMod}
\title{Server Shiny Module for \code{ibiVizEdit} that serves as a basic wrapper and enables dynamic updating of checkbox UI}
\usage{
dynamicCheckBoxInputMod(
  input,
  output,
  session,
  label = NULL,
  choices = NULL,
  selected = NULL,
  paste_char = NULL
)
}
\arguments{
\item{input, output, session}{internal parameters for {shiny}}

\item{label}{value displayed on ui for the specific checkbox}

\item{choices}{values available for selection}

\item{selected}{the default set of selections}

\item{paste_char}{optional value or vector that appends to each value defined by {choices} and {selected}}
}
\description{
Server Shiny Module for \code{ibiVizEdit} that serves as a basic wrapper and enables dynamic updating of checkbox UI
}
