% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dynamicCheckBoxInput.R
\name{dynamicCheckBoxInputModUI}
\alias{dynamicCheckBoxInputModUI}
\title{UI Shiny Module for \code{ibiVizEdit} that serves as a basic wrapper and enables dynamic updating of checkbox UI}
\usage{
dynamicCheckBoxInputModUI(id = NULL)
}
\arguments{
\item{id}{for the namespace that connects the ui and server components of the module}
}
\description{
UI Shiny Module for \code{ibiVizEdit} that serves as a basic wrapper and enables dynamic updating of checkbox UI
}
