% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{dynamicClrButtonMod}
\alias{dynamicClrButtonMod}
\title{Server-side utility for \code{ibiVizEdit} that dynamically switches actionButton UIs based on color}
\usage{
dynamicClrButtonMod(
  input,
  output,
  session,
  status_name = NULL,
  label = NULL,
  hotkey = NULL,
  hotkey_map = NULL,
  updated_label = NULL,
  default_display_name = NULL,
  button_name = "click_in",
  active_color = BUTTON_COLORS["standard"],
  inactive_color = BUTTON_COLORS["inactive"],
  updated_color = BUTTON_COLORS["warning"]
)
}
\description{
Server-side utility for \code{ibiVizEdit} that dynamically switches actionButton UIs based on color
}
