% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dynamicClrButton.R
\name{dynamicClrButtonModUI}
\alias{dynamicClrButtonModUI}
\title{UI module for \code{ibiVizEdit} that serves as a basic wrapper and enables single actionButton generation.}
\usage{
dynamicClrButtonModUI(id = NULL, inline = FALSE)
}
\arguments{
\item{id}{value for the namespace that links the server and ui components of the module}

\item{inline}{setting passed on to the ui defining whether the button should be placed in line or not}
}
\description{
UI module for \code{ibiVizEdit} that serves as a basic wrapper and enables single actionButton generation.
}
