% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dynamicClrButton.R
\name{dynamicClrButtonModUI}
\alias{dynamicClrButtonModUI}
\title{UI module for \code{ibiVizEdit} that serves as a basic wrapper and enables single actionButton generation.}
\usage{
dynamicClrButtonModUI(id = NULL, inline = FALSE)
}
\description{
UI module for \code{ibiVizEdit} that serves as a basic wrapper and enables single actionButton generation.
}
