% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{estimate_average_HR}
\alias{estimate_average_HR}
\title{Internal utility for \code{ibiVizEdit} that generates a global estimated mean HR}
\usage{
estimate_average_HR(ibi_data = NULL, ibi_col = "IBI", trim = 3)
}
\description{
Internal utility for \code{ibiVizEdit} that generates a global estimated mean HR
}
