% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{estimate_avg_respiration}
\alias{estimate_avg_respiration}
\title{Internal utility for determing average respiration jointly using PPG and IBI signals}
\usage{
estimate_avg_respiration(
  ibi_data = NULL,
  respiration_cat = NULL,
  ds = NULL,
  respiration_mapping = AVERAGE_RESPIRATION_BY_AGE,
  ibi_col = "IBI",
  time_col = "Time"
)
}
\description{
Internal utility for determing average respiration jointly using PPG and IBI signals
}
