% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{eventTriggerMod}
\alias{eventTriggerMod}
\title{Server-side utility for \code{ibiVizEdit} that enables Event Observation from action buttons inside modules}
\usage{
eventTriggerMod(
  input,
  output,
  session,
  input_id = NULL,
  trigger_items = NULL,
  trigger_values = NULL,
  trigger_object = NULL,
  trigger_id = NULL
)
}
\description{
Note that currently the logic on whether a function (func) should run needs to be contained inside that function
}
