% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{extract_ibi_editing_info}
\alias{extract_ibi_editing_info}
\title{Server side function to extract necessary information from the IBI time series}
\usage{
extract_ibi_editing_info(ibi_data, selected_points = NULL)
}
\description{
Takes in the ibi time series and extracts information used in the average, combine, and divide computations
}
