% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_impute_utils.R
\name{extract_valid_local_HP_stats}
\alias{extract_valid_local_HP_stats}
\title{Internal utility for extracting local HP mean and sd based on user-identified valid data}
\usage{
extract_valid_local_HP_stats(
  ibi_data = NULL,
  time_min = NULL,
  time_max = NULL,
  selected_points = NULL,
  input_windows = NULL,
  ibi_col = "IBI",
  time_col = "Time"
)
}
\description{
Internal utility for extracting local HP mean and sd based on user-identified valid data
}
