% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{fileButtons}
\alias{fileButtons}
\title{UI utility for \code{ibiVizEdit} that generates a series of file and director input objects}
\usage{
fileButtons(
  button_color = BUTTON_COLORS["standard"],
  heading = "Select File(s) & Working Directory"
)
}
\arguments{
\item{button_color}{hexidecimal code for the button coloring}

\item{heading}{heading info presented to the user}
}
\description{
UI utility for \code{ibiVizEdit} that generates a series of file and director input objects
}
