% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ibis.R
\name{find_ibis}
\alias{find_ibis}
\title{Integrated function for generating interbeat interval time-series.}
\usage{
find_ibis(x, ds)
}
\arguments{
\item{x}{a vector of signal values collected from a photopleythysmogram sensor. If not a vector of values must be an
object that can be safely cast to a single vector of temporally sequenced numeric values.}

\item{ds}{the desired sampling rate, in Hz, for the data used in the peak detection algorithm.}
}
\value{
Returns a list containing a \code{dataframe} of interbeat intervals and their timing and a \code{dataframe}
of diagnostic information about the peak detection algorithm optimization parameters and summary statistics.
}
\description{
\code{find_ibis} identifies the timing of heart beats in photoplethysmogram data. Input is a time-series of cardiac
data generated via recordings from photoplethysmogram sensors. The function does three main things: optimizes
bandwidth used in the peak detection algorithm, uses the peaks as proxies for the timing of heart beats, and returns
a time-series of interbeat intervals with basic diagnostic statistics.
}
