% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_utilities.R
\name{generate_base_gui_plot}
\alias{generate_base_gui_plot}
\title{Internal utility for gui plot generation}
\usage{
generate_base_gui_plot(
  ibi_data = NULL,
  color_map = NULL,
  ibi_col = "IBI",
  time_col = "Time"
)
}
\arguments{
\item{ibi_data}{is a \code{data.frame} that contains the IBI series being activel edited by the user during the
session}

\item{ibi_col}{is of type \code{character} and is the column name in the \code{ibi_data} that contains the IBI series}

\item{time_col}{is of type \code{character} and is the column name in the \code{ppg_data} that contains the time
variable}
}
\description{
\code{generate_base_gui_plot} defines the base plot that forms the basis for interactively editing IBIs
}
