% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gp_impute_core.R
\name{generate_gp_ppg_predictions}
\alias{generate_gp_ppg_predictions}
\title{Internal \code{ibiVizEdit} utility for selecting maximum a posteriori estimates of corrupted PPG data}
\usage{
generate_gp_ppg_predictions(model_outputs = NULL, gp_driver = NULL)
}
\description{
Internal \code{ibiVizEdit} utility for selecting maximum a posteriori estimates of corrupted PPG data
}
