% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{generate_heads_up_info}
\alias{generate_heads_up_info}
\title{Server-side utility for \code{ibiVizEdit} that updates summary stats displayed in text window}
\usage{
generate_heads_up_info(input, hover_id = NULL, ibi_data = NULL)
}
\description{
Server-side utility for \code{ibiVizEdit} that updates summary stats displayed in text window
}
