% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gp_impute_core.R
\name{generate_imputation_time}
\alias{generate_imputation_time}
\title{Internal utility for defining time variable over which to impute}
\usage{
generate_imputation_time(
  ppg_data = NULL,
  time_min = NULL,
  time_max = NULL,
  ds = NULL,
  time_col = "Time"
)
}
\description{
Internal utility for defining time variable over which to impute
}
