% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gp_impute_core.R
\name{generate_model_ppg_inputs}
\alias{generate_model_ppg_inputs}
\title{Internal utility for generating imputation input data}
\usage{
generate_model_ppg_inputs(
  time_min = NULL,
  time_max = NULL,
  ppg_data = NULL,
  ds = NULL,
  input_windows = NULL,
  ppg_col = "PPG",
  time_col = "Time"
)
}
\arguments{
\item{time_min}{lower boundary of the time window that defines the PPG data selected for imputation}

\item{time_max}{upper boundary of the time window that defines the PPG data selected for imputation}

\item{ppg_data}{\code{data.frame} that contains the PPG data}

\item{ds}{downsampled sampling rate in Hz}

\item{input_windows}{a \code{list} that contains the pre and post timing boundaries for the input data used in the 
imputation model}

\item{ppg_col}{name of the column in the PPG \code{data.frame} that contains the PPG signal}

\item{time_col}{name of the column in the PPG \code{data.frame} that contains timing information in seconds}
}
\value{
a set of inputs for the imputation model
}
\description{
\code{generate_model_ppg_inputs} is used to generate a \code{data.frame} that contains an appropriately down-sampled
set of "Time" and "PPG" inputs for the imputation model. These inputs represent PPG data that "surrounds" the window
selected for imputation. The size of the window is dynamically determined for each participant based on average
respiration rate (derived from \code{estimate_avg_respiration()}).
}
