% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{generate_path_messages}
\alias{generate_path_messages}
\title{Server-side utility for \code{ibiVizEdit} that outputs relevant file and directory information when submitted}
\usage{
generate_path_messages(default_text = NULL, msg_part1 = NULL, obj_name = NULL)
}
\arguments{
\item{default_text}{the default message to display}

\item{msg_part1}{the first portion of the message which remains static.}

\item{obj_name}{the object name referred to at the end of the displayed message. Typically a file or directory path}
}
\value{
a message to be displayed in the appropriate location in the {ibiVizEdit} ui
}
\description{
Server-side utility for \code{ibiVizEdit} that outputs relevant file and directory information when submitted
}
