% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gp_impute_Bayes.R
\name{gp_impute_driver}
\alias{gp_impute_driver}
\title{Internal utility that aggregates data and settings needed to trigger an imputation model run}
\usage{
gp_impute_driver(
  iter = NULL,
  warmup = NULL,
  adapt_delta = NULL,
  time_min = NULL,
  time_max = NULL,
  ppg_data = NULL,
  ibi_data = NULL,
  respiration_cat = NULL,
  ds = NULL,
  selected_ibis = NULL,
  ppg_col = "PPG",
  ibi_col = "IBI",
  time_col = "Time",
  expansion_factor = 3
)
}
\description{
Internal utility that aggregates data and settings needed to trigger an imputation model run
}
