% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graphing.R
\name{highlight_ibis}
\alias{highlight_ibis}
\title{Internal \code{ibiVizEdit} utility that enables dynamic color selection based on selected points}
\usage{
highlight_ibis(
  base_plot = NULL,
  selected_points = NULL,
  time_col = "Time",
  ibi_col = "IBI"
)
}
\arguments{
\item{base_plot}{is a \code{ggplot2} object, typically generated by \code{generate_base_gui_plot}}

\item{selected_points}{values returned from user either by clicking or using a "brush" to identify a set of IBIs for
editing}

\item{time_col}{is of type \code{character} and is the column name in the \code{ppg_data} that contains the time}

\item{ibi_col}{is of type \code{character} and is the column name in the \code{ibi_data} that contains the IBI series}
}
\description{
Internal \code{ibiVizEdit} utility that enables dynamic color selection based on selected points
}
