% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{ibi_editing_plot}
\alias{ibi_editing_plot}
\title{Server-side utility for \code{ibiVizEdit} that dynamically updates full IBI + PPG combo plots}
\usage{
ibi_editing_plot(ibi_data = DYNAMIC_DATA[["edited_ibi"]], brush_in = NULL)
}
\description{
Server-side utility for \code{ibiVizEdit} that dynamically updates full IBI + PPG combo plots
}
