% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graphing.R
\name{ibi_value_label}
\alias{ibi_value_label}
\title{Internal utility for generating dynamic text label corresponding to ibi value}
\usage{
ibi_value_label(
  base_plot = NULL,
  hover_point = NULL,
  time_col = "Time",
  ibi_col = "IBI"
)
}
\arguments{
\item{base_plot}{is a \code{ggplot2} object, typically generated by \code{generate_base_gui_plot}}

\item{hover_point}{value based on user hovering behavior - can be used to inspect specific IBI value to assist 
with editing.}

\item{time_col}{is of type \code{character} and is the column name in the \code{ppg_data} that contains the time}

\item{ibi_col}{is of type \code{character} and is the column name in the \code{ibi_data} that contains the IBI series}
}
\description{
Internal utility for generating dynamic text label corresponding to ibi value
}
