% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{idNameFields}
\alias{idNameFields}
\title{UI utility for \code{ibiVizEdit} that generates fields for ID, Study, and Editor Labels}
\usage{
idNameFields(heading = "File ID and Information:")
}
\description{
UI utility for \code{ibiVizEdit} that generates fields for ID, Study, and Editor Labels
}
