% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_utils.R
\name{out_text_dir}
\alias{out_text_dir}
\title{Shiny reactive object.}
\usage{
out_text_dir()
}
\description{
\code{out_text_dir} is an internal utility that provides a means of either printing a warning that a working
directory has not been selected or a message specifying the file path of the working directory selected.
}
