% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_utils.R
\name{out_text_time_file}
\alias{out_text_time_file}
\title{Shiny reactive object.}
\usage{
out_text_time_file()
}
\description{
\code{out_text_time_file} is an internal utility that provides a means of either printing a warning that a raw PPG
file has not been selected or a message specifying the file path of the PPG file selected.
}
