% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{ppgEditingPlots}
\alias{ppgEditingPlots}
\title{Utility for \code{ibiVizEdit} that generates main editing plots for PPG tab}
\usage{
ppgEditingPlots()
}
\description{
Utility for \code{ibiVizEdit} that generates main editing plots for PPG tab
}
