% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{ppgEditingTools}
\alias{ppgEditingTools}
\title{Utility for \code{ibiVizEdit} that generates UI components for ibi editing tab tools}
\usage{
ppgEditingTools()
}
\description{
Utility for \code{ibiVizEdit} that generates UI components for ibi editing tab tools
}
