% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{preProcessPlots}
\alias{preProcessPlots}
\title{Utility for \code{ibiVizEdit} that defines plots for pre-processing tab}
\usage{
preProcessPlots(heading = "Visualize Pre-Processed PPG Data:")
}
\description{
Utility for \code{ibiVizEdit} that defines plots for pre-processing tab
}
