% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{raise_not_in_range_integer}
\alias{raise_not_in_range_integer}
\title{Internal \code{ibiVizEdit} utility that generates a warning if an input expected to be an integer is not}
\usage{
raise_not_in_range_integer(
  input_val = NULL,
  input_name = NULL,
  lower_bound = NULL,
  upper_bound = NULL
)
}
\description{
Internal \code{ibiVizEdit} utility that generates a warning if an input expected to be an integer is not
}
