% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gp_impute_Bayes.R
\name{replace_w_imputed}
\alias{replace_w_imputed}
\title{Internal \code{ibiVizEdit} function for replacing corrupted data with imputed data}
\usage{
replace_w_imputed(
  ppg_out = NULL,
  model_outputs = NULL,
  ppg_col = "PPG",
  time_col = "Time"
)
}
\arguments{
\item{ppg_out}{PLACEHOLDER}

\item{model_outputs}{list returned from run_bayesian_gp.}

\item{ppg_col}{the column name in the PPG \code{data.frame} that contains the PPG.}

\item{time_col}{the column name in the PPG \code{data.frame} that contains the time values.}
}
\value{
the ppg_data with the imputed values overwriting the original values
}
\description{
Internal \code{ibiVizEdit} function for replacing corrupted data with imputed data
}
