% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{restore_button_action}
\alias{restore_button_action}
\title{Server side utility that takes restores all IBIs within the selected window}
\usage{
restore_button_action(
  input,
  restore_id,
  edited_data,
  original_data,
  brush_id,
  ibi_or_ppg = NULL
)
}
\description{
Server side utility that takes restores all IBIs within the selected window
}
