% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gp_impute_core.R
\name{return_stan_code}
\alias{return_stan_code}
\title{Internal \code{ibiVizEdit} utility for generating the GP imputation model in Stan code}
\usage{
return_stan_code()
}
\description{
Internal \code{ibiVizEdit} utility for generating the GP imputation model in Stan code
}
