% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_impute_Bayes.R
\name{run_bayesian_gp}
\alias{run_bayesian_gp}
\title{Internal \code{ibiVizEdit} utility for running a Bayesian Gaussian process imputation model}
\usage{
run_bayesian_gp(gp_driver, imputation_model)
}
\description{
Internal \code{ibiVizEdit} utility for running a Bayesian Gaussian process imputation model
}
