% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gp_impute_core.R
\name{save_gp_imputation_traceplots}
\alias{save_gp_imputation_traceplots}
\title{Internal \code{ibiVizEdit} utility for generating diagnostic traceplots}
\usage{
save_gp_imputation_traceplots(
  model_outputs = NULL,
  gp_driver = NULL,
  sub_id = NULL,
  secondary_id = NULL,
  study_id = NULL,
  out_dir = NULL
)
}
\arguments{
\item{model_outputs}{output object generated from the imputation run}

\item{gp_driver}{a \code{list} object used to define the imputation problem and includes hyperparameters}

\item{sub_id}{the subject id defined by the user and used by \code{ibiVizEdit} to label outputs}

\item{secondary_id}{a secondary id value defined by the user and used by \code{ibiVizEdit} to label outputs}

\item{study_id}{as study id value defined by the user and used by \code{ibiVizEdit} to label outputs}

\item{out_dir}{the internal directory where outputs are stored - created via a default process defined by the 
internal settings of \code{ibiVizEdit}}
}
\value{
saves imputation model diagnostics plots
}
\description{
Internal \code{ibiVizEdit} utility for generating diagnostic traceplots
}
