% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_impute_utils.R
\name{save_model_summary_as_text}
\alias{save_model_summary_as_text}
\title{Internal \code{ibiVizEdit} utility for generation a model output summary in .txt}
\usage{
save_model_summary_as_text(
  model_ouputs = NULL,
  gp_driver = NULL,
  sub_id = NULL,
  secondary_id = NULL,
  study_id = NULL,
  out_dir = NULL
)
}
\description{
Internal \code{ibiVizEdit} utility for generation a model output summary in .txt
}
