% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_utilities.R
\name{save_screenshot}
\alias{save_screenshot}
\title{Server side function that takes and saves screenshot in a screenshots folder for easy sharing with colleagues}
\usage{
save_screenshot(input, data, time_brush, button_id, ibi_or_ppg = NULL)
}
\description{
Server side function that takes and saves screenshot in a screenshots folder for easy sharing with colleagues
}
