% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{set_file_size_max}
\alias{set_file_size_max}
\title{Setting maximum file size}
\usage{
set_file_size_max(size = 150)
}
\description{
\code{set_file_size_max} is an internal utility that sets the maximum file size for an input data set in megabytes.
The default is set to 150 MB. If this is too low in general or for a specific use case, it can be adjusted here.
}
