% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_input_and_process.R
\name{time_center}
\alias{time_center}
\title{Internal utility for ibiVizEdit that centers PPG and IBI timing values a the start of a task file.}
\usage{
time_center(x, time_col = "Time", timing_series = NULL)
}
\arguments{
\item{x}{PPG or IBI \code{data.frame}}

\item{time_col}{the name of the column in \code{x} that contains the time data}

\item{timing_series}{the name of the \code{vector} that contains timestamps for the initial task and subsequent tasks.
Must be formatted such that the element is the start of the first task/condition, the second element is the end of
that task, the third is start of the second condition/task, the fourth is the end of that task and so on.}
}
\value{
a re-coded version of the inputs with new timing
}
\description{
\code{time_center} takes a PPG and/or IBI file and centers it at the very start of a task file. The effect is to cut
off any data prior to the initial time stamp and to set that value to 0, representing the start of the condition or
conditions of interest.
}
