% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_input_and_process.R
\name{trim_ppg_window}
\alias{trim_ppg_window}
\title{Internal utility that trims down the time range of the PPG signal used to edit the data}
\usage{
trim_ppg_window(ppg_data, timing_data, time_col = "Time")
}
\description{
Internal utility that trims down the time range of the PPG signal used to edit the data
}
