% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{turn_on_load_button}
\alias{turn_on_load_button}
\title{Server-side utility for \code{ibiVizEdit} that turns on "load" button}
\usage{
turn_on_load_button()
}
\description{
Server-side utility for \code{ibiVizEdit} that turns on "load" button
}
