% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{uneditable_button_action}
\alias{uneditable_button_action}
\title{Server side utility that marks selected points as uneditable}
\usage{
uneditable_button_action(input, ibi_data, selected_points = NULL)
}
\description{
Server side utility that marks selected points as uneditable
}
